requirements/_base.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements/_base.txt requirements/_base.in
#
anyio==3.4.0
    # via httpcore
certifi==2021.10.8
    # via
    #   httpcore
    #   httpx
charset-normalizer==2.0.9
    # via httpx
click==8.0.3
    # via -r requirements/_base.in
envyaml==1.9.210927
    # via -r requirements/_base.in
h11==0.12.0
    # via httpcore
httpcore==0.14.3
    # via httpx
httpx==0.21.1
    # via -r requirements/_base.in
idna==3.3
    # via
    #   anyio
    #   rfc3986
    #   yarl
multidict==5.2.0
    # via yarl
pydantic==1.8.2
    # via -r requirements/_base.in
pyyaml==6.0
    # via envyaml
rfc3986[idna2008]==1.5.0
    # via httpx
sniffio==1.2.0
    # via
    #   anyio
    #   httpcore
    #   httpx
typing-extensions==4.0.1
    # via pydantic
yarl==1.7.2
    # via -r requirements/_base.in
